@@ -1,16 +1,17 @@
 export const urlPre = `https://nangdinh2.herokuapp.com/`;
+export const questionNumberPractice = 50;
 // export const urlPre = `http://localhost:8000`;
 // export const urlPre = `http://localhost:8000/data`;
 
 export const testData = [
     {
-        ans: "c",
-        ch_a: "FL110",
-        ch_b: "FL100",
-        ch_c: "FL090",
-        ch_d: "FL080",
-        id: 1,
-        ques: "The MEA of Q1 in the segment between MUMGA and AC is : he MEA of Q1 in the segment between MUMGA",
+        id: 4,
+        ques: "A transmission from one station to another station in circumstances where two-way communication cannot be established but where it is believed that the called station is able to receive the transmission:",
+        ch_a: "Blind transmission",
+        ch_b: "One-way transmission",
+        ch_c: "Two-way transmission",
+        ch_d: "Radio check",
+        ans: "a",
         test: true
     },
     {
@@ -24,6 +25,16 @@ export const testData = [
       test: true
     },
     {
+        ans: "c",
+        ch_a: "FL110",
+        ch_b: "FL100",
+        ch_c: "FL090",
+        ch_d: "FL080",
+        id: 1,
+        ques: "The MEA of Q1 in the segment between MUMGA and AC is : he MEA of Q1 in the segment between MUMGA",
+        test: true
+    },
+    {
         id: 3,
         ques: "In reports for arriving aircraft, the visibility observations should be:",
         ch_a: "Representative of the landing area",
@@ -32,15 +43,6 @@ export const testData = [
         ch_d: "Representative of the whole runways.",
         ans: "c",
         test: true
-    },
-    {
-        id: 4,
-        ques: "A transmission from one station to another station in circumstances where two-way communication cannot be established but where it is believed that the called station is able to receive the transmission:",
-        ch_a: "Blind transmission",
-        ch_b: "One-way transmission",
-        ch_c: "Two-way transmission",
-        ch_d: "Radio check",
-        ans: "a",
-        test: true
     }
+    
 ]
